refactor: 알림 메시지 3초 후 사라짐 처리 로직 Notification 컴포넌트에서 수행하도록 변경
